+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `isClientLoggingEnabled` to forms app docs
 - `zoom` property to `location` form element submission data docs
+- `displayAsCurrency` to form element schema
 
 ## [0.9.8] - 2021-04-15